Separate spec file for unlistedBehaviorDescription page
