+++ src/applications/disability-benefits/all-claims/tests/pages/form0781/behaviorDescriptionPages.unit.spec.js
@@ -10,9 +10,8 @@ import { inputVaTextInput } from '@department-of-veterans-affairs/platform-testi
 import * as behaviorListPages from '../../../content/form0781/behaviorListPages';
 import { makePages } from '../../../pages/form0781/behaviorDescriptions';
 import { ALL_BEHAVIOR_CHANGE_DESCRIPTIONS } from '../../../constants';
-import * as unlistedBehaviorDescriptionPage from '../../../pages/form0781/unlistedBehaviorDescriptionPage';
 
-describe('listed behavior change descriptions', () => {
+describe('behavior change description pages', () => {
   const schemas = { ...makePages() };
   const formData = {
     workBehaviors: {
@@ -96,75 +95,5 @@ describe('listed behavior change descriptions', () => {
       userEvent.click(getByText('Submit'));
       expect(onSubmit.called).to.be.true;
     });
-  });
-});
-
-describe('unlisted behavior change description', () => {
-  const { schema, uiSchema } = unlistedBehaviorDescriptionPage;
-  it('should define a uiSchema object', () => {
-    expect(uiSchema).to.be.an('object');
-  });
-
-  it('should define a schema object', () => {
-    expect(schema).to.be.an('object');
-  });
-
-  it('Displays a text area', () => {
-    const onSubmit = Sinon.spy();
-    const { container } = render(
-      <DefinitionTester
-        schema={schema}
-        uiSchema={uiSchema}
-        data={{}}
-        definitions={{}}
-        formData={{}}
-        onSubmit={onSubmit}
-      />,
-    );
-
-    const textarea = $('va-textarea', container);
-    // Not required
-    expect(textarea.getAttribute('required')).to.eq('false');
-
-    expect(textarea.getAttribute('label')).to.eq(
-      behaviorListPages.unlistedDescriptionPageDescription,
-    );
-  });
-
-  it('should submit without entering any text', () => {
-    const onSubmit = Sinon.spy();
-
-    const { getByText } = render(
-      <DefinitionTester
-        schema={schema}
-        uiSchema={uiSchema}
-        data={{}}
-        onSubmit={onSubmit}
-      />,
-    );
-    userEvent.click(getByText('Submit'));
-    expect(onSubmit.calledOnce).to.be.true;
-  });
-
-  it('should submit if text entered', () => {
-    const onSubmit = Sinon.spy();
-
-    const { container, getByText } = render(
-      <DefinitionTester
-        schema={schema}
-        uiSchema={uiSchema}
-        data={{}}
-        onSubmit={onSubmit}
-      />,
-    );
-
-    inputVaTextInput(
-      container,
-      'Here is a description of an unlisted behavioral change',
-      'va-textarea',
-    );
-
-    userEvent.click(getByText('Submit'));
-    expect(onSubmit.called).to.be.true;
   });
 });
